requirements.txt
streamlit
openai
requests
PyGithub
